added fully resolving pmunu plaquette function; added tar ball JHW_NEW_PG_ANI.tar.gz which contains all modified files in one package;

+++ generic/layout_hyper_sl32.c
@@ -432,7 +432,7 @@ const int *get_logical_coordinate(){
 /* Map node number and index to coordinates  */
 void get_coords(int coords[], int node, size_t index){
   int mc[4];
-  int ir,kr,d,eo;
+  size_t ir,kr,d,eo;
   int k = node;
 
   /* Compute machine coordinates for node */
@@ -486,13 +486,13 @@ void get_coords(int coords[], int node, size_t index){
   /* Consistency checks for debugging */
   if((k = node_number(coords[0], coords[1], coords[2], coords[3])) 
      != node){
-    printf("get_coords: coords %d %d %d %d for node %d index %d map to wrong node %d\n",
+    printf("get_coords: coords %d %d %d %d for node %d index %ld map to wrong node %d\n",
 	   coords[0], coords[1], coords[2], coords[3], node, index, k);
     terminate(1);
   }
   if((k = node_index(coords[0], coords[1], coords[2], coords[3]))
       != index){
-    printf("get_coords: coords %d %d %d %d for node %d index %d map to wrong index %d\n",
+    printf("get_coords: coords %d %d %d %d for node %d index %ld map to wrong index %d\n",
 	   coords[0], coords[1], coords[2], coords[3], node, index, k);
     terminate(1);
   }
